Add direction preferences
User may set a preference for which direction ambiguous months are
parsed as.

For example, using 'prefer_future', the date 'june' would always refer
to the current or next June. With 'prefer_past', 'june' would always
refer to the current or previous June.

The default is 'prefer_closest', which will pick the closest (out of
past or future) with a slight preference to the future.

This should be extended to weekdays when they are supported.